writepost

+++ K-Content/src/main/java/com/example/myapp/user/commu/model/CommuFile.java
@@ -2,6 +2,7 @@ package com.example.myapp.user.commu.model;
 
 
 import com.example.myapp.user.commu.status.CommuFileDeleteYn;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -17,5 +18,5 @@ public class CommuFile {
 	    private String commuFileRegistDate; //등록일시  
 	    private String commuFileUpdateDate; //수정일시
 	    private CommuFileDeleteYn commuFileDeleteYn; // 삭제여부 Yes, No로 구분
-	    private long commuFileCommuId; // 게시글 ID (첨부파일이 있는 글번호)
+	    private int commuFileCommuId; // 게시글 ID (첨부파일이 있는 글번호)
 }
